Add execute method for ops to initiate unlocked fund withdraw

+++ src/staking.rs
@@ -1,6 +1,6 @@
 use cosmwasm_std::{StakingMsg, Response, Coin};
 
-pub fn delegate(response: Response, validator: String, amount: u64, denom: String) -> Response {
+pub fn delegate(response: Response, validator: String, amount: u128, denom: String) -> Response {
     let msg = StakingMsg::Delegate {
         validator,
         amount: Coin::new(amount.into(), denom),
@@ -8,7 +8,7 @@ pub fn delegate(response: Response, validator: String, amount: u64, denom: Strin
     response.add_message(msg)
 }
 
-pub fn redelegate(response: Response, src_validator: String, dst_validator: String, amount: u64, denom: String) -> Response {
+pub fn redelegate(response: Response, src_validator: String, dst_validator: String, amount: u128, denom: String) -> Response {
     let msg = StakingMsg::Redelegate {
         src_validator,
         dst_validator,
@@ -17,7 +17,7 @@ pub fn redelegate(response: Response, src_validator: String, dst_validator: Stri
     response.add_message(msg)
 }
 
-pub fn undelegate(response: Response, validator: String, amount: u64, denom: String) -> Response {
+pub fn undelegate(response: Response, validator: String, amount: u128, denom: String) -> Response {
     let msg = StakingMsg::Undelegate {
         validator,
         amount: Coin::new(amount.into(), denom),
